AnalysisFramework/requirements.txt
mysql-connector-python
psutil